blastoff_app/pyproject.toml
[project]
name = "blastoff_app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "zprp24z team 7", email = "lukwoj27@gmail.com"},
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "bcrypt==3.2.0",
    "cffi==1.17.1",
    "click==8.1.7",
    "cryptography==43.0.3",
    "fastapi==0.115.4",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "passlib==1.7.4",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.36",
    "starlette==0.41.2",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.0",
    "pip==24.3.1",
    "PyJWT>=2.10.1",
    "python-dotenv>=1.0.1",
    "sentry-sdk[fastapi]>=2.19.2",
    "pydantic-settings>=2.7.0",
    "redis>=5.2.1",
    "aioredis>=2.0.1",
    "fastapi-cache2>=0.2.2",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "apscheduler>=3.11.0",
    "aiosmtplib>=3.0.2",
    "requests>=2.32.3",
    "alembic>=1.14.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "pytest>=8.3.4",
    "hypothesis>=6.123.2",
    "pytest-asyncio>=0.25.1",
    "httpx>=0.28.1",
    "mkdocstrings[python]>=0.27.0",
    "mkdocstrings-python>=1.13.0",
    "mkdocs-material>=9.5.49",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "mkdocs-section-index>=0.3.9",
    "mkdocs-autorefs>=1.3.0",
]

[tool.black]
line-length = 79
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
black = "black ."
isort = "isort ."
format = {composite = ["isort", "black"]}
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
